+++ src/Support/Number.php
@@ -3,6 +3,7 @@
 namespace OpenSoutheners\ExtendedLaravel\Support;
 
 use Closure;
+use OpenSoutheners\ByteUnitConverter\ByteUnitConverter;
 
 use function OpenSoutheners\ExtendedPhp\Numbers\short_number;
 
@@ -20,5 +21,13 @@ class Number
          * Get shorter version of a big number if possible.
          */
         return fn (int|float $number): string => short_number($number);
+    }
+
+    public function toByteUnit(): Closure
+    {
+        /**
+         * Get byte unit version of a number.
+         */
+        return fn (int|string $number): ByteUnitConverter => ByteUnitConverter::new($number);
     }
 }